move find_versions_from_entites + template formatting
