+++ assistant/templates/assistant_mandates.html
@@ -59,13 +59,15 @@
 			<form method="post" action=" {% url 'mandate_change_state' %} ">
 			{% csrf_token %}
 			<td>{{ mandate.academic_year }}</td>	
-			<td><ul>
-            {% for entity in mandate.entities %}
-                    <li>
-                        {% trans entity.entity_type %} : {{ entity.acronym }}
-                    </li>
-			{% endfor %}
-            </ul></td>
+			<td>
+                <ul>
+                    {% for entity in mandate.entities %}
+                        <li>
+                            {% trans entity.entity_type %} : {{ entity.acronym }}
+                        </li>
+			        {% endfor %}
+                </ul>
+            </td>
 			<td>{{ mandate.fulltime_equivalent }}</td>
             <td>{% trans mandate.state %}</td>
 			<td align="left">
